+++ src/urh/dev/BackendHandler.py
@@ -1,10 +1,9 @@
-from subprocess import call, DEVNULL
-
-from urh import constants
 import os
 import sys
 from enum import Enum
+from subprocess import call, DEVNULL
 
+from urh import constants
 from urh.util.Logger import logger
 
 
@@ -89,7 +88,15 @@ class BackendHandler(object):
 
     def __init__(self):
 
-        self.python2_exe = constants.SETTINGS.value('python2_exe', self.__get_python2_interpreter())
+        python2_exe = constants.SETTINGS.value('python2_exe', '')
+        if not python2_exe:
+            self.__python2_exe = self.__get_python2_interpreter()
+            constants.SETTINGS.setValue("python2_exe", self.__python2_exe)
+        else:
+            self.__python2_exe = python2_exe
+
+        constants.SETTINGS.setValue("python2_exe", self.__python2_exe)
+
         self.gnuradio_install_dir = constants.SETTINGS.value('gnuradio_install_dir', "")
         self.use_gnuradio_install_dir = constants.SETTINGS.value('use_gnuradio_install_dir', os.name == "nt", bool)
 
@@ -108,6 +115,16 @@ class BackendHandler(object):
         """:type: dict[str, BackendContainer] """
 
         self.get_backends()
+
+    @property
+    def python2_exe(self):
+        return self.__python2_exe
+
+    @python2_exe.setter
+    def python2_exe(self, value):
+        if value != self.__python2_exe:
+            self.__python2_exe = value
+            constants.SETTINGS.setValue("python2_exe", value)
 
     @property
     def num_native_backends(self):
@@ -200,7 +217,7 @@ class BackendHandler(object):
                     self.gnuradio_is_installed = call('"{0}" -c "import gnuradio"'.format(self.python2_exe),
                                                       shell=True, stderr=DEVNULL) == 0
                 except OSError as e:
-                    logger.error("Could not determine GNU Radio install status. Assuming true. Error: "+str(e))
+                    logger.error("Could not determine GNU Radio install status. Assuming true. Error: " + str(e))
                     self.gnuradio_is_installed = True
             else:
                 self.gnuradio_is_installed = False
@@ -266,6 +283,14 @@ class BackendHandler(object):
             ab, rx_suprt, tx_suprt = self.__avail_backends_for_device(device_name)
             container = BackendContainer(device_name.lower(), ab, rx_suprt, tx_suprt)
             self.device_backends[device_name.lower()] = container
+
+    def get_key_from_device_display_text(self, displayed_device_name):
+        displayed_device_name = displayed_device_name.lower()
+        for key in self.DEVICE_NAMES:
+            key = key.lower()
+            if displayed_device_name.startswith(key):
+                return key
+        return None
 
     @staticmethod
     def perform_soundcard_health_check():