+++ src/urh/dev/native/lib/hackrf.pyx
@@ -1,4 +1,6 @@
-cimport chackrf
+include "config.pxi"
+
+cimport urh.dev.native.lib.chackrf as chackrf
 from libc.stdlib cimport malloc
 import time
 
@@ -39,8 +41,36 @@ cdef int _c_callback_send(chackrf.hackrf_transfer*transfer)  with gil:
 cdef chackrf.hackrf_device*_c_device
 cdef int hackrf_success = chackrf.HACKRF_SUCCESS
 
-cpdef has_multi_device_support():
-    return chackrf.HACKRF_HAS_MULTI_DEVICE != 0
+IF HACKRF_MULTI_DEVICE_SUPPORT == 1:
+    cpdef has_multi_device_support():
+        return True
+    cpdef open(str serial_number=""):
+        if not serial_number:
+            return chackrf.hackrf_open(&_c_device)
+
+        desired_serial = serial_number.encode('UTF-8')
+        c_desired_serial = <char *> desired_serial
+        return chackrf.hackrf_open_by_serial(c_desired_serial, &_c_device)
+    cpdef get_device_list():
+        init()
+        cdef chackrf.hackrf_device_list_t* device_list = chackrf.hackrf_device_list()
+
+        result = []
+        cdef int i
+        for i in range(device_list.devicecount):
+            serial_number = device_list.serial_numbers[i].decode("UTF-8")
+            result.append(serial_number)
+
+        chackrf.hackrf_device_list_free(device_list)
+        exit()
+        return result
+ELSE:
+    cpdef has_multi_device_support():
+        return False
+    cpdef open(str serial_number=""):
+        return chackrf.hackrf_open(&_c_device)
+    cpdef get_device_list():
+        return None
 
 cpdef setup(str serial):
     """
@@ -52,14 +82,6 @@ cpdef setup(str serial):
 
 cpdef init():
     return chackrf.hackrf_init()
-
-cpdef open(str serial_number=""):
-    if not chackrf.HACKRF_HAS_MULTI_DEVICE or not serial_number:
-        return chackrf.hackrf_open(&_c_device)
-
-    desired_serial = serial_number.encode('UTF-8')
-    c_desired_serial = <char *> desired_serial
-    return chackrf.hackrf_open_by_serial(c_desired_serial, &_c_device)
 
 cpdef exit():
     return chackrf.hackrf_exit()
@@ -152,20 +174,3 @@ cpdef set_amp_enable(value):
 cpdef set_baseband_filter_bandwidth(bandwidth_hz):
     time.sleep(TIMEOUT)
     return chackrf.hackrf_set_baseband_filter_bandwidth(_c_device, bandwidth_hz)
-
-cpdef get_device_list():
-    if not chackrf.HACKRF_HAS_MULTI_DEVICE:
-        return None
-
-    init()
-    cdef chackrf.hackrf_device_list_t* device_list = chackrf.hackrf_device_list()
-
-    result = []
-    cdef int i
-    for i in range(device_list.devicecount):
-        serial_number = device_list.serial_numbers[i].decode("UTF-8")
-        result.append(serial_number)
-
-    chackrf.hackrf_device_list_free(device_list)
-    exit()
-    return result